Summarizer/requirements.txt
beautifulsoup4>=4.12,<5
pytest>=7,<9
httpx>=0.27,<0.28
readability-lxml>=0.8,<0.9
markdownify>=0.12,<0.13
